+++ api/src/main/java/hr/fer/proinz/airelm/repository/ActorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActorRepository extends JpaRepository<Actor, Long> {
-    Actor findByActorID(long actorID);
+public interface ActorRepository extends JpaRepository<Actor, Integer> {
+    Actor findByActorID(Integer actorID);
 }
